requirements.txt
flask
python-dotenv
steam
zappa